--- a/benchmarks_test.go
+++ b/benchmarks_test.go
@@ -8,46 +8,21 @@ import (
 	"github.com/raditzlawliet/kdbush"
 )
 
-// Benchmark BuildIndexWith func
-func BenchmarkBuildIndexWith(b *testing.B) {
-	var points = []struct {
-		Points []kdbush.Point
-		Total  int
-	}{
-		{Points: []kdbush.Point{}, Total: 1000},
-		{Points: []kdbush.Point{}, Total: 10_000},
-		{Points: []kdbush.Point{}, Total: 100_000},
-		{Points: []kdbush.Point{}, Total: 1_000_000},
-	}
-
-	// Setup
-	for num := range points {
-		for i := 0; i < points[num].Total; i++ {
-			points[num].Points = append(points[num].Points, &kdbush.SimplePoint{rand.Float64()*24.0 + 24.0, rand.Float64()*24.0 + 24.0})
-		}
-
-	}
-
-	for _, v := range points {
-		b.Run(fmt.Sprintf("total_%d", v.Total), func(b *testing.B) {
-			for i := 0; i < b.N; i++ {
-				kdbush.NewBush().
-					BuildIndexWith(v.Points, kdbush.STANDARD_NODE_SIZE)
-			}
-		})
-	}
-}
-
 // Benchmark BuildIndex func
 func BenchmarkBuildIndex(b *testing.B) {
 	var points = []struct {
-		Points []kdbush.Point
-		Total  int
+		Points   []kdbush.Point
+		Total    int
+		NodeSize int
 	}{
-		{Points: []kdbush.Point{}, Total: 1000},
-		{Points: []kdbush.Point{}, Total: 10_000},
-		{Points: []kdbush.Point{}, Total: 100_000},
-		{Points: []kdbush.Point{}, Total: 1_000_000},
+		{Points: []kdbush.Point{}, NodeSize: kdbush.STANDARD_NODE_SIZE, Total: 1000},
+		{Points: []kdbush.Point{}, NodeSize: kdbush.STANDARD_NODE_SIZE, Total: 10_000},
+		{Points: []kdbush.Point{}, NodeSize: kdbush.STANDARD_NODE_SIZE, Total: 100_000},
+		{Points: []kdbush.Point{}, NodeSize: kdbush.STANDARD_NODE_SIZE, Total: 1_000_000},
+		{Points: []kdbush.Point{}, NodeSize: 8, Total: 1000},
+		{Points: []kdbush.Point{}, NodeSize: 8, Total: 10_000},
+		{Points: []kdbush.Point{}, NodeSize: 8, Total: 100_000},
+		{Points: []kdbush.Point{}, NodeSize: 8, Total: 1_000_000},
 	}
 
 	// Setup
@@ -59,11 +34,10 @@ func BenchmarkBuildIndex(b *testing.B) {
 	}
 
 	for _, v := range points {
-		b.Run(fmt.Sprintf("total_%d", v.Total), func(b *testing.B) {
+		b.Run(fmt.Sprintf("nodeSize_%d_total_%d", v.NodeSize, v.Total), func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
 				kdbush.NewBush().
-					Add(v.Points...).
-					BuildIndex(kdbush.STANDARD_NODE_SIZE)
+					BuildIndex(v.Points, v.NodeSize)
 			}
 		})
 	}
@@ -72,14 +46,23 @@ func BenchmarkBuildIndex(b *testing.B) {
 // Benchmark Range func
 func BenchmarkRange(b *testing.B) {
 	var points = []struct {
-		Points []kdbush.Point
-		Total  int
-		KDBush *kdbush.KDBush
+		Points   []kdbush.Point
+		Total    int
+		KDBush   *kdbush.KDBush
+		NodeSize int
 	}{
-		{Points: []kdbush.Point{}, Total: 1000},
-		{Points: []kdbush.Point{}, Total: 10_000},
-		{Points: []kdbush.Point{}, Total: 100_000},
-		{Points: []kdbush.Point{}, Total: 1_000_000},
+		{Points: []kdbush.Point{}, NodeSize: kdbush.STANDARD_NODE_SIZE, Total: 10},
+		{Points: []kdbush.Point{}, NodeSize: kdbush.STANDARD_NODE_SIZE, Total: 100},
+		{Points: []kdbush.Point{}, NodeSize: kdbush.STANDARD_NODE_SIZE, Total: 1_000},
+		{Points: []kdbush.Point{}, NodeSize: kdbush.STANDARD_NODE_SIZE, Total: 10_000},
+		{Points: []kdbush.Point{}, NodeSize: kdbush.STANDARD_NODE_SIZE, Total: 100_000},
+		{Points: []kdbush.Point{}, NodeSize: kdbush.STANDARD_NODE_SIZE, Total: 1_000_000},
+		{Points: []kdbush.Point{}, NodeSize: 8, Total: 10},
+		{Points: []kdbush.Point{}, NodeSize: 8, Total: 100},
+		{Points: []kdbush.Point{}, NodeSize: 8, Total: 1_000},
+		{Points: []kdbush.Point{}, NodeSize: 8, Total: 10_000},
+		{Points: []kdbush.Point{}, NodeSize: 8, Total: 100_000},
+		{Points: []kdbush.Point{}, NodeSize: 8, Total: 1_000_000},
 	}
 
 	// Setup
@@ -88,11 +71,11 @@ func BenchmarkRange(b *testing.B) {
 			points[num].Points = append(points[num].Points, &kdbush.SimplePoint{rand.Float64()*24.0 + 24.0, rand.Float64()*24.0 + 24.0})
 		}
 		points[num].KDBush = kdbush.NewBush().
-			BuildIndexWith(points[num].Points, kdbush.STANDARD_NODE_SIZE)
+			BuildIndex(points[num].Points, points[num].NodeSize)
 	}
 
 	for _, v := range points {
-		b.Run(fmt.Sprintf("total_%d", v.Total), func(b *testing.B) {
+		b.Run(fmt.Sprintf("nodeSize_%d_total_%d", v.NodeSize, v.Total), func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
 				v.KDBush.Range(10.0, 10, -10, -10)
 			}
@@ -103,14 +86,23 @@ func BenchmarkRange(b *testing.B) {
 // Benchmark Within func
 func BenchmarkWithin(b *testing.B) {
 	var points = []struct {
-		Points []kdbush.Point
-		Total  int
-		KDBush *kdbush.KDBush
+		Points   []kdbush.Point
+		Total    int
+		KDBush   *kdbush.KDBush
+		NodeSize int
 	}{
-		{Points: []kdbush.Point{}, Total: 1000},
-		{Points: []kdbush.Point{}, Total: 10_000},
-		{Points: []kdbush.Point{}, Total: 100_000},
-		{Points: []kdbush.Point{}, Total: 1_000_000},
+		{Points: []kdbush.Point{}, NodeSize: kdbush.STANDARD_NODE_SIZE, Total: 10},
+		{Points: []kdbush.Point{}, NodeSize: kdbush.STANDARD_NODE_SIZE, Total: 100},
+		{Points: []kdbush.Point{}, NodeSize: kdbush.STANDARD_NODE_SIZE, Total: 1_000},
+		{Points: []kdbush.Point{}, NodeSize: kdbush.STANDARD_NODE_SIZE, Total: 10_000},
+		{Points: []kdbush.Point{}, NodeSize: kdbush.STANDARD_NODE_SIZE, Total: 100_000},
+		{Points: []kdbush.Point{}, NodeSize: kdbush.STANDARD_NODE_SIZE, Total: 1_000_000},
+		{Points: []kdbush.Point{}, NodeSize: 8, Total: 10},
+		{Points: []kdbush.Point{}, NodeSize: 8, Total: 100},
+		{Points: []kdbush.Point{}, NodeSize: 8, Total: 1_000},
+		{Points: []kdbush.Point{}, NodeSize: 8, Total: 10_000},
+		{Points: []kdbush.Point{}, NodeSize: 8, Total: 100_000},
+		{Points: []kdbush.Point{}, NodeSize: 8, Total: 1_000_000},
 	}
 
 	// Setup
@@ -119,13 +111,13 @@ func BenchmarkWithin(b *testing.B) {
 			points[num].Points = append(points[num].Points, &kdbush.SimplePoint{rand.Float64()*24.0 + 24.0, rand.Float64()*24.0 + 24.0})
 		}
 		points[num].KDBush = kdbush.NewBush().
-			BuildIndexWith(points[num].Points, kdbush.STANDARD_NODE_SIZE)
+			BuildIndex(points[num].Points, points[num].NodeSize)
 	}
 
 	for _, v := range points {
-		b.Run(fmt.Sprintf("total_%d", v.Total), func(b *testing.B) {
+		b.Run(fmt.Sprintf("nodeSize_%d_total_%d", v.NodeSize, v.Total), func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
-				v.KDBush.Within(&kdbush.SimplePoint{10, 10}, 10)
+				v.KDBush.Within(10, 10, 10)
 			}
 		})
 	}