--- a/kdbush.go
+++ b/kdbush.go
@@ -6,43 +6,27 @@ const STANDARD_NODE_SIZE = 64
 
 // KDBush an instance
 type KDBush struct {
-	Points []Point // pointer
-
-	// avoid change on-the-fly
 	nodeSize int
 	ids      []int
 	coords   []float64
+	indexed  bool
 }
 
 // NewBush return a new pointer of [KDBush]
 func NewBush() *KDBush {
-	kd := KDBush{
-		Points: []Point{},
-	}
+	kd := KDBush{}
 	return &kd
 }
 
-// Add will add a new pointer point to KDBush instance. This function accept variadic param and return this pointer of KDBush
-func (kd *KDBush) Add(points ...Point) *KDBush {
-	kd.Points = append(kd.Points, points...)
-	return kd
-}
-
-// BuildIndexWith will rebuild a new kdtree index with passed []Point with nodeSize.
-// It will throw away previously process [KDBush.Add] and [KDBush.BuildIndex] and replace with the new index.
-func (kd *KDBush) BuildIndexWith(points []Point, nodeSize int) *KDBush {
-	kd.Points = points
-	return kd.BuildIndex(nodeSize)
-}
-
-// BuildIndex will rebuild a new kdtree index with saved Point from [KDBush.Add] or even latest [KDBush.BuildIndexWith] (if any) with [nodeSize].
-func (kd *KDBush) BuildIndex(nodeSize int) *KDBush {
+// BuildIndex bulding kd-tree index given list of Points
+func (kd *KDBush) BuildIndex(points []Point, nodeSize int) *KDBush {
+	kd.indexed = false
 	kd.nodeSize = nodeSize
 
-	kd.ids = make([]int, len(kd.Points))
-	kd.coords = make([]float64, 2*len(kd.Points))
+	kd.ids = make([]int, len(points))
+	kd.coords = make([]float64, 2*len(points))
 
-	for i, v := range kd.Points {
+	for i, v := range points {
 		kd.ids[i] = i
 		kd.coords[i*2] = v.GetX()
 		kd.coords[i*2+1] = v.GetY()
@@ -50,20 +34,32 @@ func (kd *KDBush) BuildIndex(nodeSize int) *KDBush {
 
 	sort(kd.ids, kd.coords, kd.nodeSize, 0, len(kd.ids)-1, 0)
 
+	kd.indexed = true
 	return kd
 }
 
-// Range will return indexes all points across [minX], [minY], [maxX], [maxY]
+// query helper struct for API Range & Within finding result
+type query struct {
+	left  int
+	right int
+	axis  int
+}
+
+// Range will return all indexes points across [minX], [minY], [maxX], [maxY]
 func (kd *KDBush) Range(minX, minY, maxX, maxY float64) []int {
-	stack := [][]int{{0, len(kd.ids) - 1, 0}}
+	if !kd.indexed {
+		return []int{}
+	}
+
+	stack := []query{{0, len(kd.ids) - 1, 0}}
 	result := []int{}
 
 	var x, y float64
 
 	for (len(stack)) > 0 {
-		left := stack[len(stack)-1][0]
-		right := stack[len(stack)-1][1]
-		axis := stack[len(stack)-1][2]
+		left := stack[len(stack)-1].left
+		right := stack[len(stack)-1].right
+		axis := stack[len(stack)-1].axis
 		stack = append(stack[:len(stack)-1], stack[len(stack):]...) // .pop()
 
 		// search linearly
@@ -90,30 +86,33 @@ func (kd *KDBush) Range(minX, minY, maxX, maxY float64) []int {
 
 		// queue search in halves that intersect the query
 		if (axis == 0 && minX <= x) || (axis != 0 && minY <= y) {
-			stack = append(stack, []int{left, m - 1, 1 - axis})
+			stack = append(stack, query{left, m - 1, 1 - axis})
 		}
 		if (axis == 0 && maxX >= x) || (axis != 0 && maxY >= y) {
-			stack = append(stack, []int{m + 1, right, 1 - axis})
+			stack = append(stack, query{m + 1, right, 1 - axis})
 		}
 	}
 
 	return result
 }
 
-// Within will return indexes all point within radius of given single [Point]
-func (kd *KDBush) Within(point Point, radius float64) []int {
-	stack := [][]int{{0, len(kd.ids) - 1, 0}}
+// Within will return all indexes points within radius of given single [Point]
+func (kd *KDBush) Within(qx, qy float64, radius float64) []int {
+	if !kd.indexed {
+		return []int{}
+	}
+
+	stack := []query{{0, len(kd.ids) - 1, 0}}
 	result := []int{}
 
 	r2 := radius * radius
 
-	qx, qy := point.GetX(), point.GetY()
 	var x, y float64
 
 	for (len(stack)) > 0 {
-		left := stack[len(stack)-1][0]
-		right := stack[len(stack)-1][1]
-		axis := stack[len(stack)-1][2]
+		left := stack[len(stack)-1].left
+		right := stack[len(stack)-1].right
+		axis := stack[len(stack)-1].axis
 		stack = append(stack[:len(stack)-1], stack[len(stack):]...) // .pop()
 
 		// search linearly
@@ -138,24 +137,36 @@ func (kd *KDBush) Within(point Point, radius float64) []int {
 
 		// queue search in halves that intersect the query
 		if (axis == 0 && qx-radius <= x) || (axis != 0 && qy-radius <= y) {
-			stack = append(stack, []int{left, m - 1, 1 - axis})
+			stack = append(stack, query{left, m - 1, 1 - axis})
 		}
 		if (axis == 0 && qx+radius >= x) || (axis != 0 && qy+radius >= y) {
-			stack = append(stack, []int{m + 1, right, 1 - axis})
+			stack = append(stack, query{m + 1, right, 1 - axis})
 		}
 	}
 
 	return result
 }
 
+//
+// Helper get private param
+//
+
+// GetNodeSize return current nodesize
 func (kd *KDBush) GetNodeSize() int {
 	return kd.nodeSize
 }
 
+// GetIndexed return all kdtree indexes
 func (kd *KDBush) GetIndexes() []int {
 	return kd.ids
 }
 
+// GetCoords return all coords
 func (kd *KDBush) GetCoords() []float64 {
 	return kd.coords
+}
+
+// Indexed return it's KDBush already indexed or not
+func (kd *KDBush) Indexed() bool {
+	return kd.indexed
 }