--- a/kdbush_test.go
+++ b/kdbush_test.go
@@ -45,24 +45,60 @@ func TestGeneration(t *testing.T) {
 		points2 = append(points2, &kdbush.SimplePoint{rand.Float64()*24.0 + 24.0, rand.Float64()*24.0 + 24.0})
 	}
 
-	bush := kdbush.NewBush().
-		BuildIndexWith(points2, kdbush.STANDARD_NODE_SIZE)
+	bush := kdbush.NewBush()
+	assert.Equal(t, bush.Indexed(), false, "should not indexed")
+	assert.Equal(t, bush.Within(0, 0, 0), []int{}, "should return empty slice of int")
+	assert.Equal(t, bush.Range(-5, 0, 5, 0), []int{}, "should return empty slice of int")
 
-	assert.ElementsMatch(t, bush.Points, points2, "they should be have same elements")
+	bush.BuildIndex(points2, kdbush.STANDARD_NODE_SIZE)
 
-	bush.Add(points2...).
-		BuildIndex(kdbush.STANDARD_NODE_SIZE)
-	points22 := append(points2, points2...)
-
-	assert.ElementsMatch(t, bush.Points, points22, "they should be have same elements")
+	assert.Equal(t, len(bush.GetIndexes()), len(points2), "they should be have same length")
+	assert.Equal(t, bush.GetNodeSize(), kdbush.STANDARD_NODE_SIZE, "nodesize should be same")
+	assert.Equal(t, bush.Indexed(), true, "should indexed")
+	assert.Equal(t, len(bush.GetIndexes()), len(points2), "indexes length should same with points")
+	assert.Equal(t, len(bush.GetCoords()), len(points2)*2, "coords length should 2x with points")
 }
 
 // Test Range func
 func TestRange(t *testing.T) {
 	bush := kdbush.NewBush().
-		BuildIndexWith(points, kdbush.STANDARD_NODE_SIZE)
+		BuildIndex(points, kdbush.STANDARD_NODE_SIZE)
 
-	assert.ElementsMatch(t, bush.Points, points, "they should be have same elements")
+	testCases := []testCase{
+		// straight line on x=0
+		{
+			[]float64{-2.1, 0.0, 2.1, 0},
+			[]int{0, 9, 10, 11, 12},
+			[]kdbush.Point{
+				points[0],
+				points[9],
+				points[10],
+				points[11],
+				points[12],
+			},
+		},
+		// straight line on x=1.0
+		{
+			[]float64{-2.1, 1.0, 2.1, 1.0},
+			[]int{2, 8, 13},
+			[]kdbush.Point{
+				points[2],
+				points[8],
+				points[13],
+			},
+		},
+	}
+
+	for _, testCase := range testCases {
+		indexes := bush.Range(testCase.Input[0], testCase.Input[1], testCase.Input[2], testCase.Input[3])
+		assert.ElementsMatch(t, indexes, testCase.Index, "they should be have same elements")
+	}
+}
+
+// Test Range func
+func TestRangeLowerNodeSize(t *testing.T) {
+	bush := kdbush.NewBush().
+		BuildIndex(points, 4)
 
 	testCases := []testCase{
 		// straight line on x=0
@@ -98,9 +134,7 @@ func TestRange(t *testing.T) {
 // Test Within func
 func TestWithin(t *testing.T) {
 	bush := kdbush.NewBush().
-		BuildIndexWith(points, kdbush.STANDARD_NODE_SIZE)
-
-	assert.ElementsMatch(t, bush.Points, points, "they should be have same elements")
+		BuildIndex(points, kdbush.STANDARD_NODE_SIZE)
 
 	testCases := []testCase{
 		// test within inner radius
@@ -135,7 +169,7 @@ func TestWithin(t *testing.T) {
 	}
 
 	for _, testCase := range testCases {
-		indexes := bush.Within(&kdbush.SimplePoint{testCase.Input[0], testCase.Input[1]}, testCase.Input[2])
+		indexes := bush.Within(testCase.Input[0], testCase.Input[1], testCase.Input[2])
 		assert.ElementsMatch(t, indexes, testCase.Index, "they should be have same elements")
 	}
 }
@@ -143,9 +177,7 @@ func TestWithin(t *testing.T) {
 // Test Within func with Lower Node Size
 func TestWithinLowNodeSize(t *testing.T) {
 	bush := kdbush.NewBush().
-		BuildIndexWith(points, 4)
-
-	assert.ElementsMatch(t, bush.Points, points, "they should be have same elements")
+		BuildIndex(points, 4)
 
 	testCases := []testCase{
 		// test within inner radius
@@ -180,7 +212,7 @@ func TestWithinLowNodeSize(t *testing.T) {
 	}
 
 	for _, testCase := range testCases {
-		indexes := bush.Within(&kdbush.SimplePoint{testCase.Input[0], testCase.Input[1]}, testCase.Input[2])
+		indexes := bush.Within(testCase.Input[0], testCase.Input[1], testCase.Input[2])
 		assert.ElementsMatch(t, indexes, testCase.Index, "they should be have same elements")
 	}
 }