Before
```sh
BenchmarkBuildIndex/total_1000-20                 654110             17228 ns/op           40960 B/op          3 allocs/op
BenchmarkBuildIndex/total_10000-20                 23164            520225 ns/op          409600 B/op          3 allocs/op
BenchmarkBuildIndex/total_100000-20                 1587           7685670 ns/op         4014080 B/op          3 allocs/op
BenchmarkBuildIndex/total_1000000-20                 134          89170069 ns/op        40017920 B/op          3 allocs/op
BenchmarkWithin/total_1000-20           61684923               182.4 ns/op           136 B/op          6 allocs/op
BenchmarkWithin/total_10000-20          50331452               239.6 ns/op           232 B/op         10 allocs/op
BenchmarkWithin/total_100000-20         39742231               300.8 ns/op           304 B/op         13 allocs/op
BenchmarkWithin/total_1000000-20        32219763               374.3 ns/op           376 B/op         16 allocs/op
```

After
```sh
BenchmarkRange/nodeSize_64_total_10-20                  839765822               14.11 ns/op            0 B/op          0 allocs/op
BenchmarkRange/nodeSize_64_total_100-20                 157284577               76.62 ns/op            0 B/op          0 allocs/op
BenchmarkRange/nodeSize_64_total_1000-20                100000000              100.4 ns/op             0 B/op          0 allocs/op
BenchmarkRange/nodeSize_64_total_10000-20               165987278               72.31 ns/op            0 B/op          0 allocs/op
BenchmarkRange/nodeSize_64_total_100000-20              125252017               97.03 ns/op            0 B/op          0 allocs/op
BenchmarkRange/nodeSize_64_total_1000000-20             97950468               125.1 ns/op             0 B/op          0 allocs/op
BenchmarkRange/nodeSize_8_total_10-20                   1000000000               8.681 ns/op           0 B/op          0 allocs/op
BenchmarkRange/nodeSize_8_total_100-20                  696466574               17.35 ns/op            0 B/op          0 allocs/op
BenchmarkRange/nodeSize_8_total_1000-20                 460818502               26.25 ns/op            0 B/op          0 allocs/op
BenchmarkRange/nodeSize_8_total_10000-20                303527074               39.56 ns/op            0 B/op          0 allocs/op
BenchmarkRange/nodeSize_8_total_100000-20               243663084               48.67 ns/op            0 B/op          0 allocs/op
BenchmarkRange/nodeSize_8_total_1000000-20              194835972               60.02 ns/op            0 B/op          0 allocs/op
BenchmarkWithin/nodeSize_64_total_10-20                 984755572               12.07 ns/op            0 B/op          0 allocs/op
BenchmarkWithin/nodeSize_64_total_100-20                176766757               67.20 ns/op            0 B/op          0 allocs/op
BenchmarkWithin/nodeSize_64_total_1000-20               134922337               89.54 ns/op            0 B/op          0 allocs/op
BenchmarkWithin/nodeSize_64_total_10000-20              192972932               62.07 ns/op            0 B/op          0 allocs/op
BenchmarkWithin/nodeSize_64_total_100000-20             151349331               79.71 ns/op            0 B/op          0 allocs/op
BenchmarkWithin/nodeSize_64_total_1000000-20            100000000              101.4 ns/op             0 B/op          0 allocs/op
BenchmarkWithin/nodeSize_8_total_10-20                  1000000000               8.625 ns/op           0 B/op          0 allocs/op
BenchmarkWithin/nodeSize_8_total_100-20                 612873906               19.54 ns/op            0 B/op          0 allocs/op
BenchmarkWithin/nodeSize_8_total_1000-20                425276205               27.61 ns/op            0 B/op          0 allocs/op
BenchmarkWithin/nodeSize_8_total_10000-20               302259685               40.02 ns/op            0 B/op          0 allocs/op
BenchmarkWithin/nodeSize_8_total_100000-20              100000000              136.8 ns/op             0 B/op          0 allocs/op
BenchmarkWithin/nodeSize_8_total_1000000-20             203591168               56.46 ns/op            0 B/op          0 allocs/op
```